refactor: DataStore 파일명 변경(#18)
