--- core/datastore/src/main/java/com/example/datastore/TokenDataStore.kt
@@ -1,39 +1,0 @@
-package com.example.datastore
-
-import android.content.Context
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.edit
-import androidx.datastore.preferences.core.stringPreferencesKey
-import androidx.datastore.preferences.preferencesDataStore
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.map
-
-private const val TOKEN_DATASTORE = "token_datastore"
-
-val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = TOKEN_DATASTORE)
-
-class TokenDataStore(private val context: Context) {
-    companion object {
-        private val TOKEN_KEY = stringPreferencesKey("auth_token")
-    }
-
-    suspend fun saveToken(token: String) {
-        context.dataStore.edit { preferences ->
-            preferences[TOKEN_KEY] = token
-        }
-    }
-
-    val token: Flow<String?> = context.dataStore.data.map { preferences ->
-        preferences[TOKEN_KEY]
-    }
-
-    suspend fun getToken(): String? = context.dataStore.data.first()[TOKEN_KEY]
-
-    suspend fun clearToken() {
-        context.dataStore.edit { preferences ->
-            preferences.remove(TOKEN_KEY)
-        }
-    }
-}
